Refactor report APIs into own files
