+++ default/methods/report/report_runner.py
@@ -1266,201 +1266,3 @@ class Report_Runner():
 
         return human_readable
 
-
-@routes.route('/api/v1/report/info/<int:report_template_id>',
-              methods = ['GET'])
-@General_permissions.grant_permission_for(['normal_user'])
-def report_info_api(report_template_id):
-    """
-    Permissions handled by Report_Runner
-    """
-
-    with sessionMaker.session_scope() as session:
-        report_runner = Report_Runner(
-            session = session,
-            member = None,
-            report_template_id = report_template_id
-        )
-
-        if len(report_runner.log["error"].keys()) >= 1:
-            return jsonify(log = report_runner.log), 400
-
-        report_runner.get_existing_report_template(report_template_id)
-
-        report_runner.validate_existing_report_id_permissions()
-
-        report_runner.log['success'] = True
-
-        return jsonify(
-            log = report_runner.log,
-            report_template = report_runner.report_template.serialize()), 200
-
-
-@routes.route('/api/v1/report/save',
-              methods = ['POST'])
-@General_permissions.grant_permission_for(['normal_user'])
-def report_save_api():
-    """
-    May or may not have an ID if it's new.
-
-    I think it's save for save to be different from running it.
-
-    metadata meaning it's data one level removed from actual report
-    ie how the report should be structured
-    see report_spec for an example
-
-    """
-    spec_list = [
-        {"report_template_id": {
-            'kind': int,
-            'required': False  # (ie for first save)
-        }
-        },
-        {'metadata': dict}
-    ]
-
-    log, input, untrusted_input = regular_input.master(request = request,
-                                                       spec_list = spec_list)
-    if len(log["error"].keys()) >= 1:
-        return jsonify(log = log), 400
-
-    with sessionMaker.session_scope() as session:
-
-        report_runner = Report_Runner(
-            session = session,
-            member = None,
-            report_template_id = input['report_template_id'],
-            metadata = input['metadata']
-        )
-
-        if len(report_runner.log["error"].keys()) >= 1:
-            return jsonify(log = report_runner.log), 400
-
-        # TODO abstract this check into a better function to share with running it
-        if input['metadata'].get('diffgram_wide_default') is True:
-
-            # TODO edge case of default wide === false (After
-            # prior being True) is not handled yet
-
-            user = User.get(session)
-            if user is None or user.is_super_admin is not True:
-                log['error']['permission'] = "'diffgram_wide_default' Invalid permission."
-                return jsonify(log = log), 400
-        else:
-            report_runner.validate_report_permissions_scope(
-                scope = input['metadata'].get('scope'),
-                project_string_id = input['metadata'].get('project_string_id'),
-            )
-
-        report_runner.save()
-
-        if len(report_runner.log["error"].keys()) >= 1:
-            return jsonify(log = report_runner.log), 400
-
-        report_runner.log['success'] = True
-
-        return jsonify(
-            log = report_runner.log,
-            report_template = report_runner.report_template.serialize()), 200
-
-
-@routes.route('/api/v1/report/run',
-              methods = ['POST'])
-@General_permissions.grant_permission_for(['normal_user'])
-def run_report_api():
-    """
-
-    Question:
-        [] Idea that we could hit the same end point,
-        and either pass a specification for the report we want
-        OR an ID??
-
-        Another option is two sperate APIs that join up
-        when they get to the "Core" part of it...
-
-        Should this also return report metadata?
-        (seems silly to have a seperate thing for that? or)...
-
-        I guess it's possible we may want to get report info without ...
-
-    Answer:
-     For now we assume we must have a saved report to run it
-     as a future optimization can look at more ways to
-     run this just in "preview" mode...
-
-    """
-
-    spec_list = [
-        {"report_template_id": {
-            'kind': int,
-            'required': False
-        }
-        },
-        {"report_template_data": {
-            'kind': dict,
-            'required': False
-        }
-        },
-        {"project_string_id": {
-            'kind': str,
-            'default': None,
-            'required': False
-        }
-        }
-    ]
-
-    log, input, untrusted_input = regular_input.master(
-        request = request,
-        spec_list = spec_list)
-
-    if len(log["error"].keys()) >= 1:
-        return jsonify(log = log), 400
-    if input.get('report_template_id') is None and input.get('report_template_data') is None:
-        log['error']['report_template'] = 'Provide report_template_id or report_template_data'
-        return jsonify(log = log), 400
-
-    with sessionMaker.session_scope() as session:
-
-        report_runner = Report_Runner(
-            session = session,
-            member = None,
-            report_template_id = input['report_template_id'],
-            report_template_data = input['report_template_data'],
-            project_string_id = input['project_string_id']
-        )
-
-        if len(report_runner.log["error"].keys()) >= 1:
-            return jsonify(log = report_runner.log), 400
-
-        report_runner.get_existing_report_template(input['report_template_id'])
-
-        """
-        For Diffgram wide reports, they only need to validate the project string id
-        BUT if it's not, then the project_string_id should match too.
-        """
-        if report_runner.report_template is not None:
-            if report_runner.report_template.diffgram_wide_default is True:
-                report_runner.validate_existing_report_id_permissions(
-                    project_string_id = input['project_string_id'])
-            else:
-                # This assume project based...
-                # this should be part of that other permission scope validation.
-                # testing but going to say http://127.0.0.1:8085/report/4
-                # with a non super user, will need to revist this for super admins too
-                if report_runner.report_template.project.project_string_id != input['project_string_id']:
-                    raise Forbidden("No access to this project.")
-
-                report_runner.validate_existing_report_id_permissions(
-                    project_string_id = input['project_string_id'])
-        else:
-            # Case where not report_template_id is provided (only report_template_data)
-            Project_permissions.user_has_project(Roles = ["admin", "Editor", "Viewer"])
-        results = report_runner.run()
-
-        if len(report_runner.log["error"].keys()) >= 1:
-            return jsonify(log = report_runner.log), 400
-
-        log['success'] = True
-        return jsonify(log = report_runner.log,
-                       report_template = report_runner.report_template.serialize(),
-                       stats = results), 200
